pyproject.toml
[tool.poetry]
name = "ezancestry"
version = "0.1.0"
description = "Easily predict and visualize genetic ancestry. Evaluate custom ancestry-informative SNP sets."
authors = ["arvkevi <arvkevi@gmail.com>"]
license = "MIT"
packages = [
    { include = "*", from = "data" },
]
readme = "README.md"

[tool.poetry.scripts]
ezancestry = "ezancestry.commands:app"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
cmake = "^3.21.2"
cyvcf2 = "^0.30.11"
loguru = "^0.5.3"
numpy = "^1.21.2"
pandas = "^1.3.2"
plotly = "^5.3.1"
scikit-learn = "^0.24.2"
snps = "^2.3.0"
streamlit = "^0.87.0"
umap-learn = "^0.5.1"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
